remove the templating
