+++ pipelines/datasets/br_fgv_igp/flows.py
@@ -61,10 +61,31 @@ Flows for br_fgv_igp
 from prefect import Parameter, case
 from prefect.run_configs import KubernetesRun
 from prefect.storage import GCS
-from pipelines.constants import constants
-from pipelines.utils.crawler_fgv_igp.tasks import hello_task
 
+from pipelines.constants import constants
+from pipelines.datasets import every_month
+from pipelines.datasets.br_fgv_igp.tasks import crawler_fgv, hello_task
 from pipelines.utils.decorators import Flow
+
+with Flow(
+    name="IGP-DI mensal",
+    code_owners=[
+        "Mauricio Fagundes",
+    ],
+) as fgv_igp_flow:
+    # Parameters
+    INDICE = Parameter("indice", default="IGP12_IGPDI12")
+    PERIODO = Parameter("periodo", default="mensal")
+    dataset_id = Parameter("dataset_id", default="br_fgv_igp")
+    table_id = Parameter("table_id")
+
+    crawler_task = crawler_fgv(INDICE)
+
+
+fgv_igp_flow.storage = GCS(constants.GCS_FLOWS_BUCKET.value)
+fgv_igp_flow.run_config = KubernetesRun(image=constants.DOCKER_IMAGE.value)
+fgv_igp_flow.schedule = every_month
+
 
 with Flow(
     name="my_flow",
