+++ pipelines/utils/__init__.py
@@ -8,4 +8,5 @@ Prefect flows for basedosdados project
 from pipelines.utils.crawler_ibge_inflacao.flows import *
 from pipelines.utils.dump_to_gcs.flows import *
 from pipelines.utils.execute_dbt_model.flows import *
-from pipelines.utils.crawler_fgv_igp.flows import *
+
+# from pipelines.utils.crawler_fgv_igp.flows import *
