--- pipelines/utils/crawler_fgv_igp/constants.py
@@ -1,42 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Constant values for the utils projects
-"""
-
-
-###############################################################################
-#
-# Esse é um arquivo onde podem ser declaratas constantes que serão usadas
-# pelo projeto crawler_fgv_igp.
-#
-# Por ser um arquivo opcional, pode ser removido sem prejuízo ao funcionamento
-# do projeto, caos não esteja em uso.
-#
-# Para declarar constantes, basta fazer conforme o exemplo abaixo:
-#
-# ```
-# class constants(Enum):
-#     """
-#     Constant values for the crawler_fgv_igp project
-#     """
-#     FOO = "bar"
-# ```
-#
-# Para usá-las, basta fazer conforme o exemplo abaixo:
-#
-# ```py
-# from pipelines.utils.crawler_fgv_igp.constants import constants
-# print(constants.FOO.value)
-# ```
-#
-###############################################################################
-
-from enum import Enum
-
-
-class constants(Enum):  # pylint: disable=c0103
-    """
-    Constant values for the crawler_fgv_igp project
-    """
-
-    FOO = "bar"
