--- pipelines/utils/crawler_fgv_igp/flows.py
@@ -1,87 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Flows for crawler_fgv_igp
-"""
-
-###############################################################################
-#
-# Aqui é onde devem ser definidos os flows do projeto.
-# Cada flow representa uma sequência de passos que serão executados
-# em ordem.
-#
-# Mais informações sobre flows podem ser encontradas na documentação do
-# Prefect: https://docs.prefect.io/core/concepts/flows.html
-#
-# De modo a manter consistência na codebase, todo o código escrito passará
-# pelo pylint. Todos os warnings e erros devem ser corrigidos.
-#
-# Existem diversas maneiras de declarar flows. No entanto, a maneira mais
-# conveniente e recomendada pela documentação é usar a API funcional.
-# Em essência, isso implica simplesmente na chamada de funções, passando
-# os parâmetros necessários para a execução em cada uma delas.
-#
-# Também, após a definição de um flow, para o adequado funcionamento, é
-# mandatório configurar alguns parâmetros dele, os quais são:
-# - storage: onde esse flow está armazenado. No caso, o storage é o
-#   próprio módulo Python que contém o flow. Sendo assim, deve-se
-#   configurar o storage como o pipelines.utils
-# - run_config: para o caso de execução em cluster Kubernetes, que é
-#   provavelmente o caso, é necessário configurar o run_config com a
-#   imagem Docker que será usada para executar o flow. Assim sendo,
-#   basta usar constants.DOCKER_IMAGE.value, que é automaticamente
-#   gerado.
-# - schedule (opcional): para o caso de execução em intervalos regulares,
-#   deve-se utilizar algum dos schedules definidos em schedules.py
-#
-# Um exemplo de flow, considerando todos os pontos acima, é o seguinte:
-#
-# -----------------------------------------------------------------------------
-# from prefect import task
-# from prefect import Flow
-# from prefect.run_configs import KubernetesRun
-# from prefect.storage import GCS
-# from pipelines.constants import constants
-# from my_tasks import my_task, another_task
-# from my_schedules import some_schedule
-#
-# with Flow("my_flow") as flow:
-#     a = my_task(param1=1, param2=2)
-#     b = another_task(a, param3=3)
-#
-# flow.storage = GCS(constants.GCS_FLOWS_BUCKET.value)
-# flow.run_config = KubernetesRun(image=constants.DOCKER_IMAGE.value)
-# flow.schedule = some_schedule
-# -----------------------------------------------------------------------------
-#
-# Abaixo segue um código para exemplificação, que pode ser removido.
-#
-###############################################################################
-
-
-from prefect import Parameter
-from prefect.run_configs import KubernetesRun
-from prefect.storage import GCS
-
-from pipelines.constants import constants
-from pipelines.utils.crawler_fgv_igp.schedules import every_month
-from pipelines.utils.crawler_fgv_igp.tasks import crawler_fgv
-from pipelines.utils.decorators import Flow
-
-with Flow(
-    name="BD Template: crawler FGV IGP",
-    code_owners=[
-        "Mauricio Fagundes",
-    ],
-) as fgv_igp_flow:
-    # Parameters
-    INDICE = Parameter("indice", default="IGP12_IGPDI12")
-    PERIODO = Parameter("periodo")
-    dataset_id = Parameter("dataset_id", default="br_fgv_igp")
-    table_id = Parameter("table_id")
-
-    crawler_task = crawler_fgv(INDICE)
-
-
-fgv_igp_flow.storage = GCS(constants.GCS_FLOWS_BUCKET.value)
-fgv_igp_flow.run_config = KubernetesRun(image=constants.DOCKER_IMAGE.value)
-fgv_igp_flow.schedule = every_month
