--- pipelines/utils/crawler_fgv_igp/utils.py
@@ -1,31 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-General purpose functions for the crawler_fgv_igp project
-"""
-
-###############################################################################
-#
-# Esse é um arquivo onde podem ser declaratas funções que serão usadas
-# pelo projeto crawler_fgv_igp.
-#
-# Por ser um arquivo opcional, pode ser removido sem prejuízo ao funcionamento
-# do projeto, caos não esteja em uso.
-#
-# Para declarar funções, basta fazer em código Python comum, como abaixo:
-#
-# ```
-# def foo():
-#     """
-#     Function foo
-#     """
-#     print("foo")
-# ```
-#
-# Para usá-las, basta fazer conforme o exemplo abaixo:
-#
-# ```py
-# from pipelines.utils.crawler_fgv_igp.utils import foo
-# foo()
-# ```
-#
-###############################################################################
